add UI class, show demo ImGUI window

+++ src/core/Image.cpp
@@ -3,7 +3,7 @@
 Image::Image(int32 height)
 	: m_height(height)
 {
-	m_width = m_height * aspectRatio;
+	m_width = m_height * int32(aspectRatio);
 }
 
 Image::Image(int32 height, int32 width)
